util: Fix compilation with GCC
GCC (at least version 14.2.1) recognizes isdigit and isspace as built-in
functions, and our versions don't match the types of standard versions
which take and return ints so GCC issues -Wbuiltin-declaration-mismatch
warning (and error with -Werror).

There is no real problem because our versions are static and we know
what we're calling, but an easy way to suppress the error is to avoid
the standard library name.